src/agents/calculator/requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
httpx>=0.27.0
a2a-sdk[http-server]>=0.3.0
pydantic>=2.10.0
langchain-core>=0.3.0
langgraph>=0.2.0
databricks-langchain>=0.1.0
